fix(hikyuu_cpp): 修改NormZScore中nsigma参数类型为double

+++ hikyuu_cpp/hikyuu/trade_sys/multifactor/normalize/NormZScore.cpp
@@ -43,7 +43,7 @@ PriceList NormZScore::normalize(const PriceList& data) {
     }
 
     bool outExtreme = getParam<bool>("out-extreme");
-    int nsigma = getParam<int>("nsigma");
+    double nsigma = getParam<double>("nsigma");
     bool recursive = getParam<bool>("recursive");
 
     price_t sum = 0.0;
